Create README.md
Brazil has several official lotteries made available by the government. Here we will analyze just one lottery called “Lotofácil” and through it show mathematically and computationally the two lottery problems:

First Problem: is it possible to have a minimum number of bets that guarantee the bettor receiving at least one of the secondary prizes?

Second Problem: is it possible to identify any lottery addiction? If so, is it possible to generate a strategy that significantly increases the bettor's result?

Follow the development of this project at: https://github.com/elisama/brazilian-lotteries.git